updated vessel width extraction code
if the video is not the first one you are analyzing for the exp, you no longer have to input redundant info. the background subtraction and vessel width rois coordinates are extracted from previous .mat files. now this saves vessel width of multiple rois instead of averaging across all of them automatically